+++ src/api/java/appeng/api/definitions/IItems.java
@@ -83,6 +83,8 @@ public interface IItems {
 
 	IItemDefinition cellCreative();
 
+	IItemDefinition fluidCellCreative();
+
 	IItemDefinition viewCell();
 
 	IItemDefinition cell1k();